refactor: cleanup unused components
