--- src/components/card/SummaryCard.module.css
@@ -1,52 +1,0 @@
-
-.cardGroup {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, 320px);
-    grid-gap: var(--spacers-dp16);
-}
-
-.cardGroupHeader {
-    font-size: 22px;
-    padding: 12px 0;
-    color: var(--colors-grey900);
-    line-height: 40px;
-}
-
-.cardWrapper {
-    display: flex;
-    flex-direction: column;
-    padding: 16px;
-    width: 320px;
-    height: 100%;
-}
-
-.cardIcon svg {
-    width: 32px;
-    height: 32px;
-}
-
-.cardHeader {
-    padding: 12px 0px;
-    color: var(--colors-grey900);
-    font-size: 16px;
-    font-weight: 500;
-    line-height: 20px;
-}
-
-.cardContent {
-    font-size: 14px;
-    color: var(--colors-grey900);
-    padding: 0;
-    margin: 0;
-    line-height: 20px;
-}
-
-.cardActions {
-    /* margin-auto pushes this to the bottom (extends margin to occupy extra space) */
-    margin-top: auto;
-    padding-top: 12px;
-}
-
-.cardActions a {
-    text-decoration: none;
-}
