--- src/components/card/index.tsx
@@ -1,1 +1,0 @@
-export * from "./SummaryCard";
