--- src/components/lazy/DefaultLazyLoad.module.css
@@ -1,4 +1,0 @@
-
-.loadingSpinner {
-    margin: var(--spacers-dp8) auto;
-}
