--- src/components/lazy/DefaultLazyLoad.tsx
@@ -1,9 +1,0 @@
-import { CircularLoader } from "@dhis2/ui";
-import React, { Suspense } from "react";
-import styles from "./DefaultLazyLoad.module.css";
-
-export const DefaultLazyLoad = ({ element }) => (
-    <Suspense fallback={<CircularLoader className={styles.loadingSpinner} />}>
-        {element}
-    </Suspense>
-);
